backend/requirements.txt
flask
flask-cors
opencv-python
livekit-api
PyJWT
